refactor(desktop): split kanban page models into focused hooks
Extract kanban board, dialogs, and session-start flow hooks; move page model types into a shared file; restore contextual agent modal fallback copy; and add focused tests for new session-start helpers.
