+++ apps/desktop/src/pages/use-kanban-board-model.ts
@@ -1,0 +1,77 @@
+import type { RunSummary, TaskCard } from "@openducktor/contracts";
+import { mapToKanbanColumns } from "@openducktor/core";
+import { useMemo } from "react";
+import type { AgentSessionState } from "@/types/agent-orchestrator";
+import type { KanbanPageContentModel } from "./kanban-page-model-types";
+
+const ACTIVE_SESSION_STATUS = new Set<AgentSessionState["status"]>(["starting", "running"]);
+
+type UseKanbanBoardModelArgs = {
+  tasks: TaskCard[];
+  runs: RunSummary[];
+  sessions: AgentSessionState[];
+  onOpenDetails: (taskId: string) => void;
+  onDelegate: (taskId: string) => void;
+  onPlan: (taskId: string, action: "set_spec" | "set_plan") => void;
+  onBuild: (taskId: string) => void;
+  onHumanApprove: (taskId: string) => void;
+  onHumanRequestChanges: (taskId: string) => void;
+};
+
+export function useKanbanBoardModel({
+  tasks,
+  runs,
+  sessions,
+  onOpenDetails,
+  onDelegate,
+  onPlan,
+  onBuild,
+  onHumanApprove,
+  onHumanRequestChanges,
+}: UseKanbanBoardModelArgs): KanbanPageContentModel {
+  const columns = useMemo(() => mapToKanbanColumns(tasks), [tasks]);
+
+  const runStateByTaskId = useMemo(
+    () => new Map(runs.map((run) => [run.taskId, run.state])),
+    [runs],
+  );
+
+  const activeSessionsByTaskId = useMemo(() => {
+    const sessionsByTaskId = new Map<string, AgentSessionState[]>();
+    for (const session of sessions) {
+      if (!ACTIVE_SESSION_STATUS.has(session.status)) {
+        continue;
+      }
+
+      const existing = sessionsByTaskId.get(session.taskId);
+      if (existing) {
+        existing.push(session);
+      } else {
+        sessionsByTaskId.set(session.taskId, [session]);
+      }
+    }
+
+    for (const taskSessions of sessionsByTaskId.values()) {
+      taskSessions.sort((left, right) => {
+        if (left.status !== right.status) {
+          return left.status === "running" ? -1 : 1;
+        }
+        return right.startedAt.localeCompare(left.startedAt);
+      });
+    }
+
+    return sessionsByTaskId;
+  }, [sessions]);
+
+  return {
+    columns,
+    runStateByTaskId,
+    activeSessionsByTaskId,
+    onOpenDetails,
+    onDelegate,
+    onPlan,
+    onBuild,
+    onHumanApprove,
+    onHumanRequestChanges,
+  };
+}
